Landing component almost fully functionnal. Next stp is to add redirect and error messages
