+++ frontend/src/App.js
@@ -2,14 +2,13 @@ import React, { Component } from 'react';
 import {
   BrowserRouter as Router,
   Switch,
-  Route,
-  Link
+  Route
 } from "react-router-dom";
 import './App.css';
 import './css/text.css';
 import UserProfile from "./components/userProfile.component"
 import CardMedia  from './components/card.js';
-import landingpage from './components/landingPage.component';
+import landingpage from './components/landing/landing.component';
 import NavBar from './components/navbar/navbar.component';
 
 class App extends Component {
